% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_factors.R
\name{optimal_factors}
\alias{optimal_factors}
\title{number factors function}
\usage{
optimal_factors(data, kmax)
}
\arguments{
\item{data}{a matrix X.}

\item{kmax}{the maximum number of factors}
}
\value{
a list with the optimal factors.
}
\description{
number factors function
}
\examples{
data("optimal_x")
optimal_factor <- optimal_factors(optimal_x,kmax=8)
}
