% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_factor.R
\name{optimal_number_factors}
\alias{optimal_number_factors}
\title{optimal number of factors function}
\usage{
optimal_number_factors(x, kmax)
}
\arguments{
\item{x}{a matrix x.}

\item{kmax}{the maximum number of factors}
}
\value{
a list with the optimal factors.
}
\description{
optimal number of factors function
}
\examples{
data("optimal_x")
optimal_factor <- optimal_number_factors(x=optimal_x,kmax=8)
}
