% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_x.R
\docType{data}
\name{optimal_x}
\alias{optimal_x}
\title{Optimal x}
\format{A matrix with the most predictive terms.

\describe{
\item{x}{The matrix with 4 components.}
}}
\usage{
optimal_x
}
\description{
A simple \code{matrix} containing the optimal words selected by Elastic Net from 1992:01 through 2018:11.
}
\keyword{datasets}
