% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_nowcast.R
\name{text_nowcast}
\alias{text_nowcast}
\title{text nowcast}
\usage{
text_nowcast(x, y, intercept)
}
\arguments{
\item{x}{the input matrix x. It should have 1 observation more that y.}

\item{y}{the response variable}

\item{intercept}{TRUE for include intercept in the forecast equation.}
}
\value{
the nowcast h=0 for the variable y.
}
\description{
text nowcast
}
\examples{
set.seed(1)
data("stock_data")
data("news_data")
y=as.matrix(stock_data[,2])
w=as.matrix(stock_data[,3])
data("news_data")
data("optimal_factors")
pc=optimal_factors
z=cbind(w,pc)
t=length(y)
ncsts=text_nowcast(z,y[1:(t-1)],TRUE)
}
