% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_idf.R
\name{tf_idf}
\alias{tf_idf}
\title{tf-idf function}
\usage{
tf_idf(x)
}
\arguments{
\item{x}{a input matrix x of terms counting.}
}
\value{
a list with the terms tf-idf and the terms tf-idf in descending order.
}
\description{
tf-idf function
}
\examples{
data("news_data")
X=as.matrix(news_data[,2:ncol(news_data)])
tf_idf_terms = tf_idf(X)
}
