% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_factors.R
\docType{data}
\name{optimal_factors}
\alias{optimal_factors}
\title{Optimal Factors}
\format{A vector with 1 component.

\describe{
\item{optimal fators x}{The vector of factor.}
}}
\usage{
optimal_factors
}
\description{
A simple vector containing the Optimal factors  select by optimal_number_factors function.
}
\keyword{datasets}
